+++ client/src/components/board/CheckDisplay.js
