+++ client/src/components/menus/MainMenuSolo.js
